Rearrange the project layout, add pytest, add ci.sh
