+++ .github/workflows/python.yaml
@@ -13,6 +13,7 @@ jobs:
     name: Check Python
     runs-on: "${{ matrix.on }}"
     strategy:
+      fail-fast: false
       matrix:
         python-version: [3.7, 3.8, 3.9]
         on: [ubuntu-latest, windows-latest]
@@ -28,22 +29,24 @@ jobs:
     - name: Install CI dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install flake8 black mypy
+        pip install black flake8 mypy psutil pytest
 
     - name: Lint with flake8
+      shell: bash
       run: |
-        cd ptd_client_server
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --ignore E203 --exit-zero --max-line-length=127 --statistics
+        ./ptd_client_server/ci.sh flake8
 
     - name: Check code formatting with black
+      shell: bash
       run: |
-        cd ptd_client_server
-        black --check --diff .
+        ./ptd_client_server/ci.sh black
 
     - name: Check typing with mypy
+      shell: bash
       run: |
-        cd ptd_client_server
-        mypy --allow-redefinition --strict --pretty .
+        ./ptd_client_server/ci.sh mypy
+
+    - name: Run unit tests wit pytest
+      shell: bash
+      run: |
+        ./ptd_client_server/ci.sh pytest
