+++ ptd_client_server/lib/common.py
